Enchance vformat
